LayoutTests/js/dom/dom-as-prototype-assignment-exception-expected.txt

CONSOLE MESSAGE: Deprecated attempt to access property 'id' on a non-Element object.
CONSOLE MESSAGE: Deprecated attempt to access property 'id' on a non-Element object.
Tests to make sure we throw when triggering a custom property with a mismatched this

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testObject.id is undefined.
PASS testObject.id="foo" threw exception TypeError: The Element.id setter can only be used on instances of Element.
PASS testObject.length threw exception TypeError: The NodeList.length getter can only be used on instances of NodeList.
PASS div.id is 'test'
PASS div.hasOwnProperty('id') is false
PASS div.__proto__.id is undefined.
PASS successfullyParsed is true

TEST COMPLETE
